+++ scripts/delayed.js
@@ -2,6 +2,7 @@
 import { sampleRUM } from './lib-franklin.js';
 
 import { sendAnalyticsPageLoadedEvent } from './adobeDataLayer.js';
+import { addScript, instance } from './utils.js';
 
 // Core Web Vitals RUM collection
 sampleRUM('cwv');
@@ -9,3 +10,8 @@ sampleRUM('cwv');
 // add more delayed functionality here
 
 sendAnalyticsPageLoadedEvent();
+
+addScript('https://consent.cookiebot.com/uc.js', { culture: 'en', cbid: '4a55b566-7010-4633-9b03-7ba7735be0b6' }, 'defer');
+
+if (instance === 'prod') addScript('https://assets.adobedtm.com/8a93f8486ba4/5492896ad67e/launch-b1f76be4d2ee.min.js', {}, 'defer');
+else addScript('https://assets.adobedtm.com/8a93f8486ba4/5492896ad67e/launch-3e7065dd10db-staging.min.js', {}, 'defer');
