+++ styles/lazy-styles.css
@@ -1,7 +1,3 @@
 /* below the fold CSS goes here */
 
 @import url('https://fonts.googleapis.com/css2?family=Roboto:wght@400;700&display=swap');
-
-/* load components style */
-@import url("/components/column/column.css");
-@import url("/components/review/review.css");
